requirements.txt
azure-servicebus
azure-mgmt-servicebus
openai
httpx
